update 29/10 (#22)
* Provost sprites and code cleanup (#4409)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request
Gives the provost uniforms their own sprites as only the Enforcer used
the provost uniform sprite, and all others used MP sprites which
contrasted with the provost jackets/armour.
Also removes a lot of duplicate code where things were subtyped and
really didn't need to be.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Having all the duplicate entries, especially in the preset lists, is
really clunky and useless. The only things that were being used were
specially named uniforms and suits, which is unnecessary.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
imageadd: Added sprites for provost senior and marshal uniforms.
spellcheck: Gave unique names to the provost armour subtypes.
balance: Prevented most provost armour from being able to fit inside
bags.
balance: Returned provost armour to the same slowdowns as MP armour.
code: Removed a lot of duplicate code in Provost presets and
uniform/suits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

* Automatic changelog for PR #4409 [ci skip]

* Fix for incendiary OB spread. (#4791)

# About the pull request

Followup to #4645.

![image](https://github.com/cmss13-devs/cmss13/assets/4447185/905bb1d9-ef4b-4963-9632-881a7da1bf8c)


# Explain why it's good for the game


![image](https://github.com/cmss13-devs/cmss13/assets/4447185/edd79640-fb51-47b9-baa0-dfd9ca62b86b)

# Changelog
:cl:
fix: Incendiary OB once again spreads to intended size.
/:cl:

* Automatic changelog for PR #4791 [ci skip]

* Remove Direct Hit Limb Penalty against Xenos (#4788)

# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

## Simple version

Selecting a body part on doll at bottom right applies an inherent
accuracy debuff to shots (an hand is harder to hit than chest).

This penalty still applied to sprite-click attacks on Xenomorphs,
despite the localized damage having no effect.

This means that you'd have to constantly reset your aimed part to chest
every time you try to shoot a Xeno, or would suffer a "free" penalty.

## Detailled version

Old flow went as follow:
 * Get base Hit Chance and Roll
* If it's only a near hit, or target's not the shot clicked target, the
shot targeted limb is reset to a random one
 * Hit chance receives a penalty based on aimed limb
 * We proceed checking hit chance vs roll

Because of the above problem this is changed to:

* Get base Hit Chance and Roll
* If it's a near hit (including limb penalty), a near xeno hit
(excluding limb penalty), or not the sprite-clicked target, aimed limb
is reset to a random one
* Else if the target is a xeno and it was a proper direct hit, there is
no penalty -- functionally similar to a chest shot
* Else Finally, apply per-limb penalty as usual
* We proceed checking hit chance vs roll

# Explain why it's good for the game
Avoids an invisible noobtrap debuff for just about everyone in the game
that isn't in the know - and the neccessity to constantly reset
targeting to chest for medics.

# Testing Photographs and Procedure
The problem was tested by a player on discord -- the fix i tested just
was spawning stuff and shooting up the place observing RNG misses. I did
not take detailed statistics over time to confirm.


# Changelog
:cl:
fix: Sprite-click shots onto Xenos are no longer affected by
limb-targeting penalty, because it was an accuracy debuff when there is
no inherent benefit to targeting Xeno limbs.
/:cl:

* Automatic changelog for PR #4788 [ci skip]

* Automatic changelog compile [ci skip]

---------

Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>
Co-authored-by: cm13-github <128137806+cm13-github@users.noreply.github.com>
Co-authored-by: Segrain <Segrain@users.noreply.github.com>
Co-authored-by: fira <loyauflorian@gmail.com>
Co-authored-by: Changelogs <action@github.com>